fix: write the per-entity linetype scale as a real, not an integer
The "ltscale" column is declared OFTReal and DRW_Entity::ltypeScale is a
double, but addEntity() passed it to setInteger(), whose parameter is an int.
The conversion happened silently at the call, so every value was truncated
towards zero.

Linetype scales are almost always fractional — 0.5, 0.35, 0.25 are the common
ones in a German AEC drawing — so they all landed in the GeoPackage as 0. A
zero scale is not "unset", it is a different answer, and dashed lines came out
using the layer's pattern length rather than the entity's own.

Use setDouble(), which is what the neighbouring linewidth write already does.

+++ src/app/dwg/qgsdwgimporter.cpp
@@ -1501,7 +1501,11 @@ void QgsDwgImporter::addEntity( OGRFeatureDefnH dfn, OGRFeatureH f, const DRW_En
   setString( dfn, f, u"color"_s, colorString( data.color, data.color24, data.transparency, layer ) );
   setInteger( dfn, f, u"lweight"_s, DRW_LW_Conv::lineWidth2dxfInt( data.lWeight ) );
   setDouble( dfn, f, u"linewidth"_s, lineWidth( data.lWeight, layer ) );
-  setInteger( dfn, f, u"ltscale"_s, data.ltypeScale );
+  // ltscale is an OFTReal column and DRW_Entity::ltypeScale is a double.
+  // setInteger() takes an int, so routing it there narrowed every scale to its
+  // integer part: the usual sub-unity values (0.5, 0.35, 0.25) all became 0 and
+  // dashes were drawn at the layer's scale instead of the entity's.
+  setDouble( dfn, f, u"ltscale"_s, data.ltypeScale );
   SETINTEGER( visible );
 }
